dani_web_page/requirements.txt
flask==3.0.0
tradingeconomics==3.2.0
python-dotenv==1.0.0 